fix scalar

+++ scalar/src/lib.rs
@@ -1,7 +1,7 @@
 pub fn sum(left: u8, right: u8) -> u8 {
     left + right
 }
-pub fn diff(a: i32, b: i32) -> i32 {
+pub fn diff(a: i16, b: i16) -> i16 {
 a-b
 }
